# src/NileTechno.API/openapi/niletechno-openapi.yaml
openapi: 3.0.3
info:
  title: NileTechno API
  description: |
    Frontend API Reference for NileTechno E-Commerce API.
    JSON fields use camelCase. Public product/catalog data only includes active ERP products/groups.
  version: 1.0.0
servers:
  - url: https://my-back-end-jayl.onrender.com
    description: Production
  - url: http://localhost:5080
    description: Local Development
...
security:
  - {}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token. Expires after 60 minutes.
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean }
...
    CartItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: string }
        quantity: { type: integer }
        color: { type: string, nullable: true }
        size: { type: string, nullable: true }
    CartSyncRequest:
      type: object
      required: [items]
      properties:
        userId: { type: string, description: يُتجاهل لغير الأدمن — الهوية تُؤخذ من التوكن. الأدمن فقط يستطيع تحديد حساب آخر. }
        items:
          type: array
          items: { $ref: '#/components/schemas/CartItemInput' }
    CartProduct:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        title: { type: string }
        price: { type: number, format: double }
        image: { type: string }
        stock: { type: number, format: double }
        groupId: { type: string }
        category: { type: string }
    CartItem:
      type: object
      description: عنصر السلة — id هو رقم صف جدول Ec_Cart (integer وليس GUID)
      properties:
        id: { type: integer, format: int32 }
        productId: { type: string, description: كود المنتج (EC_Products.Code) }
        quantity: { type: integer }
        color: { type: string, nullable: true }
        size: { type: string, nullable: true }
        product: { $ref: '#/components/schemas/CartProduct' }
    CartUpsertRequest:
      allOf:
        - type: object
          properties:
            userId: { type: string, description: يُتجاهل لغير الأدمن — الهوية تُؤخذ من التوكن }
        - $ref: '#/components/schemas/CartItemInput'
...
    Address:
      type: object
      properties:
        id: { type: string, format: uuid }
        userId: { type: string }
        label: { type: string }
        governorate: { type: string }
        details: { type: string }
        latitude: { type: number, format: double, nullable: true }
        longitude: { type: number, format: double, nullable: true }
        isDefault: { type: boolean }
        createdAt: { type: string, format: date-time }
    AddressInput:
      type: object
      required: [label, governorate, details, isDefault]
      properties:
        userId: { type: string, description: يُتجاهل لغير الأدمن — الهوية من التوكن }
        id:
          type: string
          format: uuid
          nullable: true
          example: null
          description: |
            اختياري. اتركه null او احذفه لانشاء عنوان جديد.
            قيمته فقط عند التعديل ويجب ان تكون id عنوان موجود فعلًا لنفس الـ userId؛
            id غير موجود يرجع 404 "العنوان غير موجود".
        label: { type: string }
        governorate: { type: string }
        details: { type: string }
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
        isDefault: { type: boolean }
    AddressSyncRequest:
      type: object
      required: [addresses]
      properties:
        userId: { type: string, description: يُتجاهل لغير الأدمن — الهوية من التوكن }
        addresses:
          type: array
          items: { $ref: '#/components/schemas/AddressInput' }
...
    AbandonedCartItem:
      type: object
      required: [name, price, quantity]
      properties:
        name: { type: string }
        price: { type: number, format: double }
        quantity: { type: integer }
        image: { type: string }
    AbandonedCartRequest:
      type: object
      required: [customerName, governorate, total, items]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string }
        customerName: { type: string }
        customerPhone: { type: string }
        customerEmail: { type: string, format: email }
        governorate: { type: string }
        total: { type: number, format: double }
        items:
          type: array
          items: { $ref: '#/components/schemas/AbandonedCartItem' }
...
paths:
...
  /api/cart:
    get:
      tags: [Cart]
      summary: Get cart
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: userId, in: query, required: false, schema: { type: string }, description: للأدمن فقط — لغير الأدمن تتجاهل السيرفر القيمة وتستخدم هوية التوكن }
      responses:
        '200':
          description: Cart items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CartItem' }

    post:
      tags: [Cart]
      summary: Upsert cart item
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CartUpsertRequest' }
      responses:
        '200':
          description: Upserted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '400':
          description: ناقص userId/productId أو كود المنتج غير موجود أو موقوف (Status=0)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }

    delete:
      tags: [Cart]
      summary: Clear cart
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: userId, in: query, required: false, schema: { type: string }, description: للأدمن فقط — لغير الأدمن تتجاهل السيرفر القيمة وتستخدم هوية التوكن }
      responses:
        '200':
          description: Cart cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  removed: { type: integer }
  /api/cart/sync:
    post:
      tags: [Cart]
      summary: Replace cart contents
      security: [{ bearerAuth: [] }]
      description: يستبدل كل صفوف Ec_Cart الخاصة باليوزر بمحتوى الطلب (الحذف ثم Upsert).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CartSyncRequest' }
      responses:
        '200':
          description: Synchronized
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  replaced: { type: integer, description: عدد الصفوف التي تم ادراجها أو تحديثها }
                  removed: { type: integer, description: عدد صفوف Ec_Cart التي تم حذفها }
        '400':
          description: userId مفقود
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
  /api/cart/{itemId}:
    delete:
      tags: [Cart]
      summary: Delete cart item
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: itemId, in: path, required: true, schema: { type: integer, format: int32 }, description: رقم صف Ec_Cart (id العائد من GET /api/cart) }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '404':
          description: العنصر غير موجود
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
...
  /api/addresses:
    get:
      tags: [Addresses]
      summary: List user addresses
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: userId, in: query, required: false, schema: { type: string }, description: للأدمن فقط — لغير الأدمن تتجاهل السيرفر القيمة وتستخدم هوية التوكن }
      responses:
        '200':
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Address' }

    post:
      tags: [Addresses]
      summary: Create or update address
      security: [{ bearerAuth: [] }]
      description: |
        انشاء عنوان جديد: اترك `id` بقيمة null او احذفه من الـ body.
        تعديل عنوان: أرسل `id` عنوان موجود فعلًا لنفس الـ userId، غير ذلك يرجع 404.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddressInput' }
            example:
              userId: "37"
              id: null
              label: "Home"
              governorate: "Gharbia"
              details: "Tanta, El Bahr Street, Building 15, Apartment 4"
              latitude: 30.7865
              longitude: 31.0004
              isDefault: false
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Address' }
  /api/addresses/sync:
    post:
      tags: [Addresses]
      summary: Replace all user addresses
      security: [{ bearerAuth: [] }]
      description: |
        يستبدل القائمة بالكامل: أي عنوان محفوظ غير موجود في `addresses` يتم حذفه.
        بعث `id: null` للعنوان الجديد و`id` الحقيقي للتعديل، وابعث القائمة كاملة.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddressSyncRequest' }
      responses:
        '200':
          description: Synchronized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
  /api/addresses/{id}:
    delete:
      tags: [Addresses]
      summary: Delete address
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: userId, in: query, required: false, schema: { type: string }, description: للأدمن فقط — لغير الأدمن تتجاهل السيرفر القيمة وتستخدم هوية التوكن }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
...
  /api/abandoned-carts:
    post:
      tags: [Abandoned Carts]
      summary: Create or update abandoned cart
      description: "يبقى متاحًا للزوار بدون توكن، والمفتاح يأتي من `userId` أو `id` في الـ body (uid أو رقم موبايل أو cart_<ts>). أما لو الطلب مصحوبًا بتوكن فيُفرض المفتاح على id حساب التوكن، فلا يمكن كتابة سلة متروكة على حساب آخر."
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AbandonedCartRequest' }
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }

    get:
      tags: [Abandoned Carts]
      summary: List latest abandoned carts
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Latest 500 abandoned carts
          content:
            application/json:
              schema:
                type: array
                items: { type: object, additionalProperties: true }
        '403': { description: Admin required }
  /api/abandoned-carts/{id}:
    delete:
      tags: [Abandoned Carts]
      summary: Delete abandoned cart
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Deleted }
        '403': { description: Admin required }
...